+++ seatbid.go
@@ -15,7 +15,7 @@ import (
 type SeatBid struct {
 	Bids  []Bid           `json:"bid"`             // Array of bid objects; each realtes to an imp, if exchange supported can have many bid objects.
 	Seat  string          `json:"seat,omitempty"`  // ID of the bidder seat optional string ID of the bidder seat on whose behalf this bid is made.
-	Group int             `json:"group,omitempty"` // '1' means impression must be won-lost as a group; default is '0'.
+	Group int8            `json:"group,omitempty"` // '1' means impression must be won-lost as a group; default is '0'.
 	Ext   json.RawMessage `json:"ext,omitempty"`
 }
